requirements.txt
langchain
huggingface_hub
mysql.connector 